Changed: Moved threading logic for GPW/LRP from initializers to runners
By moving the logic of starting threads from the initializer to the runner
we are able to make standalone applications which start GPW and LRP processing
threads.

CHANGELOG
Changed: Moved threading logic for GPW/LRP from initializers to runners. Allows for making standalone application which start GPW and LRP processing threads.
