+++ core/core/src/main/java/org/visallo/core/model/longRunningProcess/LongRunningProcessRunner.java
@@ -14,6 +14,7 @@ import org.visallo.core.status.StatusServer;
 import org.visallo.core.status.model.LongRunningProcessRunnerStatus;
 import org.visallo.core.status.model.ProcessStatus;
 import org.visallo.core.user.User;
+import org.visallo.core.util.StoppableRunnable;
 import org.visallo.core.util.VisalloLogger;
 import org.visallo.core.util.VisalloLoggerFactory;
 
@@ -57,7 +58,8 @@ public class LongRunningProcessRunner extends WorkerBase {
         LongRunningWorkerPrepareData workerPrepareData = new LongRunningWorkerPrepareData(
                 map,
                 this.user,
-                InjectHelper.getInjector());
+                InjectHelper.getInjector()
+        );
         for (LongRunningProcessWorker worker : InjectHelper.getInjectedServices(LongRunningProcessWorker.class, configuration)) {
             try {
                 LOGGER.info("preparing: %s", worker.getClass().getName());
@@ -136,5 +138,47 @@ public class LongRunningProcessRunner extends WorkerBase {
     @Inject
     public void setConfiguration(Configuration configuration) {
         this.configuration = configuration;
+    }
+
+    public static List<StoppableRunnable> startThreaded(int threadCount, Configuration config) {
+        List<StoppableRunnable> stoppables = new ArrayList<>();
+
+        LOGGER.info("Starting LongRunningProcessRunners on %d threads", threadCount);
+        for (int i = 0; i < threadCount; i++) {
+            StoppableRunnable stoppable = new StoppableRunnable() {
+                private LongRunningProcessRunner longRunningProcessRunner = null;
+
+                @Override
+                public void run() {
+                    try {
+                        longRunningProcessRunner = InjectHelper.getInstance(LongRunningProcessRunner.class);
+                        longRunningProcessRunner.prepare(config.toMap());
+                        longRunningProcessRunner.run();
+                    } catch (Exception ex) {
+                        LOGGER.error("Failed running LongRunningProcessRunner", ex);
+                    }
+                }
+
+                @Override
+                public void stop() {
+                    try {
+                        if (longRunningProcessRunner != null) {
+                            LOGGER.debug("Stopping LongRunningProcessRunner");
+                            longRunningProcessRunner.stop();
+                        }
+                    } catch (Exception ex) {
+                        LOGGER.error("Failed stopping LongRunningProcessRunner", ex);
+                    }
+                }
+            };
+            stoppables.add(stoppable);
+            Thread t = new Thread(stoppable);
+            t.setName("long-running-process-runner-" + t.getId());
+            t.setDaemon(true);
+            LOGGER.debug("Starting LongRunningProcessRunner thread: %s", t.getName());
+            t.start();
+        }
+
+        return stoppables;
     }
 }
